resolve warnings

+++ src/common/workgroup.hpp
@@ -128,12 +128,13 @@ __attribute__((always_inline)) inline void wg_dft(T* loc, T* loc_twiddles, const
            * Load data from the column corresponsing to the sub_batch_being computed,
            * in a transposed fashion, viewing each column as N x M Matrix.
            */
-          transfer_strided<T, detail::pad::DO_PAD, FactWiN, 0>(priv, loc, 2 * max_num_batches_in_local_mem,
-                                                               2 * sub_batch_num, M, column, 1, wi_in_fft * FactWiN,
-                                                               BankLinesPerPad);
+          transfer_strided<T, detail::pad::DO_PAD, FactWiN, 0>(
+              priv, loc, 2 * max_num_batches_in_local_mem, 2 * sub_batch_num, static_cast<std::size_t>(M),
+              static_cast<std::size_t>(column), 1L, static_cast<std::size_t>(wi_in_fft * FactWiN), BankLinesPerPad);
         } else {
-          transfer_strided<T, detail::pad::DO_PAD, FactWiN, 0>(priv, loc, 1, 0, 2 * M, 2 * column, 1,
-                                                               wi_in_fft * FactWiN, BankLinesPerPad);
+          transfer_strided<T, detail::pad::DO_PAD, FactWiN, 0>(
+              priv, loc, 1L, 0L, static_cast<std::size_t>(2 * M), static_cast<std::size_t>(2 * column), 1L,
+              static_cast<std::size_t>(wi_in_fft * FactWiN), BankLinesPerPad);
         }
       }
       sg_dft<Dir, FactWiN, FactSgN>(priv, sg, loc_twiddles + (2 * M));
@@ -144,12 +145,14 @@ __attribute__((always_inline)) inline void wg_dft(T* loc, T* loc_twiddles, const
            * in a transposed fashion, viewing each column as N x M Matrix, given the result from
            * sg_dft is also transposed in the registers.
            */
-          transfer_strided<T, detail::pad::DO_PAD, FactWiN, 1>(priv, loc, 2 * max_num_batches_in_local_mem,
-                                                               2 * sub_batch_num, M, column, FactSgN, wi_in_fft,
-                                                               BankLinesPerPad);
+          transfer_strided<T, detail::pad::DO_PAD, FactWiN, 1>(
+              priv, loc, 2 * max_num_batches_in_local_mem, 2 * sub_batch_num, static_cast<std::size_t>(M),
+              static_cast<std::size_t>(column), static_cast<std::size_t>(FactSgN), static_cast<std::size_t>(wi_in_fft),
+              BankLinesPerPad);
         } else {
-          transfer_strided<T, detail::pad::DO_PAD, FactWiN, 1>(priv, loc, 1, 0, 2 * M, 2 * column, FactSgN, wi_in_fft,
-                                                               BankLinesPerPad);
+          transfer_strided<T, detail::pad::DO_PAD, FactWiN, 1>(
+              priv, loc, 1L, 0L, static_cast<std::size_t>(2 * M), static_cast<std::size_t>(2 * column),
+              static_cast<std::size_t>(FactSgN), static_cast<std::size_t>(wi_in_fft), BankLinesPerPad);
         }
       }
     }
@@ -197,9 +200,9 @@ __attribute__((always_inline)) inline void wg_dft(T* loc, T* loc_twiddles, const
           /**
            * Load FactWiM contiguous elements per column corresponding to the sub batch being processed.
            */
-          transfer_strided<T, detail::pad::DO_PAD, FactWiM, 0>(priv, loc, 2 * max_num_batches_in_local_mem,
-                                                               2 * sub_batch_num, 1, row * M, 1, wi_in_fft * FactWiM,
-                                                               BankLinesPerPad);
+          transfer_strided<T, detail::pad::DO_PAD, FactWiM, 0>(
+              priv, loc, 2 * max_num_batches_in_local_mem, 2 * sub_batch_num, 1L, static_cast<std::size_t>(row * M), 1L,
+              static_cast<std::size_t>(wi_in_fft * FactWiM), BankLinesPerPad);
         } else {
           local2private<2 * FactWiM, detail::pad::DO_PAD, BankLinesPerPad>(
               loc, priv, static_cast<std::size_t>(wi_in_fft), static_cast<std::size_t>(2 * FactWiM),
@@ -230,9 +233,9 @@ __attribute__((always_inline)) inline void wg_dft(T* loc, T* loc_twiddles, const
            * Store back FactWiM contiguous elements per column corresponding to the sub batch being processed,
            * un-transposing the transposed result obtained from sg_dft
            */
-          transfer_strided<T, detail::pad::DO_PAD, FactWiM, 1>(priv, loc, 2 * max_num_batches_in_local_mem,
-                                                               2 * sub_batch_num, 1, M * row, FactSgN, wi_in_fft,
-                                                               BankLinesPerPad);
+          transfer_strided<T, detail::pad::DO_PAD, FactWiM, 1>(
+              priv, loc, 2 * max_num_batches_in_local_mem, 2 * sub_batch_num, 1L, static_cast<std::size_t>(M * row),
+              static_cast<std::size_t>(FactSgN), static_cast<std::size_t>(wi_in_fft), BankLinesPerPad);
         } else {
           store_transposed<2 * FactWiM, detail::pad::DO_PAD, BankLinesPerPad>(
               priv, loc, static_cast<std::size_t>(wi_in_fft), static_cast<std::size_t>(FactSgM),
